Correctly store the win percentage (Flugkiller)
The player_win_perc view calculates this correctly as DECIMAL(5,2), but
the players table was storing it as INT. This made the tournament point
calculations for the Win Rate category incorrect, since it relies on the
ratio of this value to the maximum across players taken from the view.
Hence a player could get more or less points for the category than they
should depending on how the decimal value was rounded to integer. We now
store it as DECIMAL(5,2) in players as well.

To fix this in the live tournament, I stopped the script via `touch
taildb.stop`, made a backup of the tournaments DB with mysqldump, and
then ran the following via the mysql client:

ALTER TABLE players MODIFY COLUMN win_perc DECIMAL(5,2);

After restarting taildb, the all player scores were fixed since they're
recalculated with each update.

+++ database-tables.sql
@@ -43,7 +43,7 @@ CREATE TABLE IF NOT EXISTS players (
   highest_score INT,
   lowest_turncount_win INT,
   fastest_win INT,
-  win_perc INT,
+  win_perc DECIMAL(5,2),
   piety INT,
   banner_score INT,
   exploration INT,
